+++ client/.gitignore
@@ -24,3 +24,4 @@ yarn-debug.log*
 yarn-error.log*
 
 note.txt
+test.js
